f-analytics - Move duplicate methods to shared mapper module + + rearranged some files + used short import paths
