+++ packages/services/f-analytics/src/plugins/analytics.plugin.js
@@ -1,35 +1,19 @@
-import analyticsModule from '../store/analytics.module';
-import AnalyticService from './lib/analytics.service';
+import AnalyticService from '@/services/analytics.service';
+import analyticsModule from '@/store/analytics.module';
 import {
     PUSH_PLATFORM_DATA,
     PUSH_EVENT
-} from '../store/mutation-types';
+} from '@/store/mutation-types';
+import { mapServersidePlatformData } from '@/services/analytics.mapper';
 
-const defaults = require('./defaults');
+const defaultOptions = require('../../defaultOptions');
 
-const getCookie = (name, req) => {
-    if (req && req.headers && req.headers.cookie) {
-        const value = `; ${req.headers.cookie}`;
-        const parts = value.split(`; ${name}=`);
-        if (parts.length === 2) {
-            return parts.pop().split(';').shift();
-        }
-    }
-
-    return undefined;
-};
-
-const mapServersideValues = (store, req, options) => {
+const prepareServersideValues = (store, req, options) => {
     // Only available serverside
     if (typeof (window) === 'undefined') {
         const platformData = { ...store.state[`${options.namespace}`].platformData };
 
-        if (process.env.justEatEnvironment) platformData.environment = process.env.justEatEnvironment;
-        if (process.env.FEATURE_VERSION) platformData.version = process.env.FEATURE_VERSION;
-        if (process.env.INSTANCE_POSITION) platformData.instancePosition = process.env.INSTANCE_POSITION;
-
-        const userPercent = getCookie('je-user_percentage', req);
-        if (userPercent) platformData.jeUserPercentage = userPercent;
+        mapServersidePlatformData(platformData, req);
 
         store.dispatch(`${options.namespace}/${PUSH_PLATFORM_DATA}`, platformData);
     }
@@ -45,6 +29,7 @@ const preparePageTags = options => {
     // Only add tags if clientside and if not already added
     if (typeof (window) !== 'undefined' && !window.dataLayer) {
         const queryString = options.auth ? `&gtm_auth=${options.auth}&gtm_preview=${options.preview}&gtm_cookies_win=${options.cookiesWin}` : '';
+
         // See : https://developers.google.com/tag-manager/quickstart
         const headJsGtmTag = `(function (w, d, s, l, i) {
                                     w[l] = w[l] || [];
@@ -79,7 +64,7 @@ const preparePageTags = options => {
 
 export default ({ store, req }, inject, _options) => {
     const options = {
-        ...defaults,
+        ...defaultOptions,
         ..._options
     };
 
@@ -87,7 +72,7 @@ export default ({ store, req }, inject, _options) => {
 
     registerStoreModule(store, options);
 
-    mapServersideValues(store, req, options);
+    prepareServersideValues(store, req, options);
 
     const service = new AnalyticService(store, req, options);
