+++ packages/services/f-analytics/src/services/_tests/analytics.service.test.js
@@ -1,4 +1,4 @@
-import AnalyticService from '@/plugins/lib/analytics.service';
+import AnalyticService from '@/services/analytics.service';
 import {
     createStore,
     newEvent,
